--- framework/arcane-framework/src/main/scala/services/metrics/datadog/base/DataDogConfiguration.scala
@@ -1,18 +1,0 @@
-package com.sneaksanddata.arcane.framework
-package services.metrics.datadog.base
-
-import com.timgroup.statsd.StatsDClient
-
-
-/**
- * A trait that represents a DataDog configuration.
- */
-trait DataDogConfiguration:
-  
-  /**
-   * Builds a DataDog client.
-   *
-   * @return The DataDog client.
-   */
-  def build: StatsDClient
-
